ByteArray implemented
