+++ src/lib/CLValue2/Key.ts
@@ -1,33 +1,6 @@
 import { CLType, CLValue } from './Abstract';
 import { URef } from './URef';
-
-export class AccountHashType extends CLType {
-  toString(): string {
-    return 'URef';
-  }
-}
-
-/** A cryptographic public key. */
-export class AccountHash extends CLValue {
-  v: Uint8Array;
-  /**
-   * Constructs a new `AccountHash`.
-   *
-   * @param v The bytes constituting the public key.
-   */
-  constructor(v: Uint8Array) {
-    super();
-    this.v = v;
-  }
-
-  clType(): CLType {
-    return new AccountHashType();
-  }
-
-  value(): Uint8Array {
-    return this.v;
-  }
-}
+import { AccountHash } from "./AccountHash";
 
 export enum KeyVariant {
   Account,
